rename internal/slices -> slicex tp avoid name conflict with std go slices package
